RF + ENH: Moved stuff out of the __init__ into a module.
In the module, I have now implemented calls to web_eval as a way to call code
on the server, instead of just functions.
